test(e2e): split tests into separate app files
- update tests to work with most recent app updates
- create test-user token with a testrepo for use with projects
